Refactor FormikFileInput to always upload using storage context.
