+++ src/core/ui/upload/FileUpload/FileUploadButton.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from 'react-i18next';
 import 'react-image-crop/dist/ReactCrop.css';
 import UploadButton from '../../button/UploadButton';
 import { StorageConfig } from '../../../../domain/storage/StorageBucket/useStorageConfig';
-import { useUploadFileMutation, useUploadFileOnReferenceMutation } from '../../../apollo/generated/apollo-hooks';
+import { useUploadFileMutation } from '../../../apollo/generated/apollo-hooks';
 import { useNotification } from '../../notifications/useNotification';
 
 interface FileUploadProps {
@@ -16,19 +16,12 @@ interface FileUploadProps {
 
 const bytesInMegabyte = Math.pow(1024, 2);
 
-const FileUploadButton: FC<FileUploadProps> = ({ onUpload, referenceID, storageConfig }) => {
+const FileUploadButton: FC<FileUploadProps> = ({ onUpload, storageConfig }) => {
   const { t } = useTranslation();
   const notify = useNotification();
 
   const acceptedFileTypes = storageConfig.allowedMimeTypes.join(',');
   const maxFileSizeMb = storageConfig.maxFileSize ? storageConfig.maxFileSize / bytesInMegabyte : 0;
-
-  const [uploadFileOnReference, { loading: loadingOnReference }] = useUploadFileOnReferenceMutation({
-    onCompleted: data => {
-      notify(t('components.file-upload.file-upload-success'), 'success');
-      onUpload?.(data.uploadFileOnReference.uri);
-    },
-  });
 
   const [uploadFile, { loading: loadingOnStorageBucket }] = useUploadFileMutation({
     onCompleted: data => {
@@ -36,7 +29,7 @@ const FileUploadButton: FC<FileUploadProps> = ({ onUpload, referenceID, storageC
       onUpload?.(data.uploadFileOnStorageBucket);
     },
   });
-  const loading = loadingOnReference || loadingOnStorageBucket;
+  const loading = loadingOnStorageBucket;
 
   const handleSubmit = async (selectedFile: File) => {
     if (!selectedFile) return;
@@ -45,25 +38,14 @@ const FileUploadButton: FC<FileUploadProps> = ({ onUpload, referenceID, storageC
       notify(t('components.file-upload.file-size-error', { limit: maxFileSizeMb }), 'error');
       return;
     }
-    if (referenceID) {
-      await uploadFileOnReference({
-        variables: {
-          file: selectedFile,
-          uploadData: {
-            referenceID,
-          },
+    await uploadFile({
+      variables: {
+        file: selectedFile,
+        uploadData: {
+          storageBucketId: storageConfig.storageBucketId,
         },
-      });
-    } else {
-      await uploadFile({
-        variables: {
-          file: selectedFile,
-          uploadData: {
-            storageBucketId: storageConfig.storageBucketId,
-          },
-        },
-      });
-    }
+      },
+    });
   };
 
   return (
